pyproject.toml
[project]
name = "aves-colombia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cloudinary>=1.44.0",
    "django>=5.2.2",
    "django-environ>=0.12.0",
    "django-jazzmin>=3.0.1",
    "pillow>=11.2.1",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=5.2.0",
]
